requirements.txt
pyyaml>=6.0.2
pillow>=10.4.0
scipy>=1.10.1
pandas>=2.0.3
plotly>=6.0.0
streamlit>=1.40.1